requirements.txt
pandas
faiss-cpu
sentence-transformers
openai
transformers
streamlit
